Overhauled the view sorting mechanism.
The mechanism did not work well, and on top of that it was incomprehensible to me. I rewrote the whole algorithm. Now it is clear and works decently.
Note that it requires objects to be configured coherently to provide a coherent order. Incoherences such as big rocks crossing each other will result in buggy ordering.
Unfortunately this is not 100% preventable with the current implementation, but configuring objects with coherence should be fine. Usage will tell if the current implementation is solid enough.