Fixed bug where entity could not find next angle.
The overtake mechanism of getNextAngle sometimes caused the solution to contain 1 step, which is invalid.
This case corresponds to the entity overtaking its destination. From the pathfinder's perspective, it is fine that the entity keeps following its destination, as there is no next step in the path.
This case causes the entity to frenetically change direction while located on its destination. Functionally this is bad, but it's not the pathfinder's responsibility to handle this. It will be the entity AI's responsibility.